+++ src/main/java/site/hellishmods/reignitedutilites/lib/tileentities/CreativeMillTile.java
@@ -10,7 +10,7 @@ public class CreativeMillTile extends GPProducingTile {
     }
 
     @Override
-    public int getGPOutput() {return 10000;}
+    public Float getGPOutput() {return 10000f;} // TODO: config options for generation amount
     @Override
     public boolean canGenerate(BlockPos pos, World world) {return true;}
 }
